--- WallPlanter/manifest.json
@@ -1,22 +1,0 @@
-﻿{
-  "Name": "Wall Planter",
-  "Author": "aedenthorn",
-  "Version": "0.1.1",
-  "Description": "Wall Planter.",
-  "UniqueID": "aedenthorn.WallPlanter",
-  "EntryDll": "WallPlanter.dll",
-  "MinimumApiVersion": "3.13.0",
-  "ModUpdater": {
-    "Repository": "StardewValleyMods",
-    "User": "aedenthorn",
-    "Directory": "_releases",
-    "ModFolder": "WallPlanter"
-  },
-  "UpdateKeys": [ "Nexus:11339" ],
-  "Dependencies": [
-    {
-      "UniqueID": "Platonymous.ModUpdater",
-      "IsRequired": false
-    }
-  ]
-}
